Added leave

+++ app/Http/Controllers/SRController.php
@@ -8,6 +8,7 @@ use App\Models\DistributionOrder;
 use App\Models\Network;
 use App\Models\PurchaseOrder;
 use App\Models\Client;
+use App\Models\District;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
@@ -27,18 +28,20 @@ class SRController extends Controller
     }
     public function dataTable()
     {
-        $query = Client::where('type',2);//DSR
+        $query = Client::where('type',2);//SR
         return DataTables::eloquent($query)
             ->addIndexColumn()
             ->addColumn('action', function(Client $client) {
                 $btn = '';
-                if(auth()->user()->can('dsr_edit')){
-                    $btn  .='<a href="'.route('sr.edit',['sr'=>$client->id]).'" class="btn btn-primary bg-gradient-primary btn-sm btn-edit"><i class="fa fa-edit"></i></a>';
-                }
-                if(auth()->user()->can('dsr_delete')) {
-                    $btn .= ' <a role="button" data-id="' . $client->id . '" class="btn btn-danger btn-sm btn-delete"><i class="fa fa-trash"></i></a>';
-                }
+                $btn  .='<a href="'.route('sr.edit',['sr'=>$client->id]).'" class="btn btn-primary bg-gradient-primary btn-sm btn-edit"><i class="fa fa-edit"></i></a>';
+                $btn .= ' <a role="button" data-id="' . $client->id . '" class="btn btn-danger btn-sm btn-delete"><i class="fa fa-trash"></i></a>';
                 return $btn;
+            })
+            ->addColumn('district_name', function(Client $client) {
+                return $client->district->name_eng ?? '';
+            })
+            ->addColumn('thana_name', function(Client $client) {
+                return $client->thana->name_eng ?? '';
             })
             ->rawColumns(['action'])
             ->toJson();
@@ -50,10 +53,8 @@ class SRController extends Controller
      */
     public function create()
     {
-        if (!auth()->user()->hasPermissionTo('dsr_create')) {
-            abort(403, 'Unauthorized');
-        }
-        return view('settings.sr.create');
+        $districts = District::where('status',1)->get();
+        return view('settings.sr.create',compact('districts'));
     }
 
     /**
@@ -61,10 +62,6 @@ class SRController extends Controller
      */
     public function store(Request $request)
     {
-        if (!auth()->user()->hasPermissionTo('dsr_create')) {
-            abort(403, 'Unauthorized');
-        }
-
         // Validate the request data
         $validatedData = $request->validate([
             'name' =>[
@@ -78,12 +75,13 @@ class SRController extends Controller
                     ->where('type',2)
             ],
             'email' =>[
-                'nullable','max:255',
+                'required','max:255',
                 Rule::unique('clients')
                     ->where('type',2)
             ],
+            'district' =>['required'],
+            'thana' =>['required'],
             'address' => 'nullable|string|max:255', // Make 'address' nullable
-            'opening_balance' => 'required|numeric|min:0',
             'status' => 'required|boolean', // Ensure 'status' is boolean
         ]);
 
@@ -93,6 +91,12 @@ class SRController extends Controller
         try {
             // Create a new Client record in the database
             $validatedData['type'] = 2;
+            $validatedData['district_id'] = $validatedData['district'];
+            $validatedData['thana_id'] = $validatedData['thana'];
+
+            unset($validatedData['district']);
+            unset($validatedData['thana']);
+
             $sr = Client::create($validatedData);
             // Create/Update User
             $user = User::where('role','SR')->where('client_id',$sr->id)->first();
@@ -105,7 +109,7 @@ class SRController extends Controller
             $user->username  = $request->mobile_no;
             $user->email = $request->email;
             $user->mobile_no = $request->mobile_no;
-            $user->password = bcrypt($request->password);
+            $user->password = bcrypt('12345678');
             $user->save();
 
             // Commit the transaction
@@ -129,15 +133,13 @@ class SRController extends Controller
      */
     public function edit(Client $sr)
     {
-        if (!auth()->user()->hasPermissionTo('dsr_edit')) {
-            abort(403, 'Unauthorized');
-        }
         if ($sr->type != 2) {
             abort(404);
         }
+        $districts = District::where('status',1)->get();
         try {
             // If the Client exists, display the edit view
-            return view('settings.sr.edit', compact('sr'));
+            return view('settings.sr.edit', compact('sr','districts'));
         } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             // Handle the case where the Client is not found
             return redirect()->route('sr.index')->with('error', 'SR not found: '.$e->getMessage());
@@ -149,9 +151,6 @@ class SRController extends Controller
      */
     public function update(Request $request, Client $sr)
     {
-        if (!auth()->user()->hasPermissionTo('dsr_edit')) {
-            abort(403, 'Unauthorized');
-        }
         if ($sr->type != 2) {
             abort(404);
         }
@@ -170,13 +169,14 @@ class SRController extends Controller
                     ->ignore($sr)
             ],
             'email' =>[
-                'nullable','max:255',
+                'required','max:255',
                 Rule::unique('clients')
                     ->where('type',2)
                     ->ignore($sr)
             ],
             'address' => 'nullable|string|max:255', // Make 'address' nullable
-            'opening_balance' => 'required|numeric|min:0',
+            'district' =>['required'],
+            'thana' =>['required'],
             'status' => 'required|boolean', // Ensure 'status' is boolean
         ]);
 
@@ -186,6 +186,12 @@ class SRController extends Controller
         try {
             // Update the Client record in the database
             $validatedData['type'] = 2;
+            $validatedData['district_id'] = $validatedData['district'];
+            $validatedData['thana_id'] = $validatedData['thana'];
+
+            unset($validatedData['district']);
+            unset($validatedData['thana']);
+
             $sr->update($validatedData);
 
             // Create/Update User
@@ -199,7 +205,6 @@ class SRController extends Controller
             $user->username  = $request->mobile_no;
             $user->email = $request->email;
             $user->mobile_no = $request->mobile_no;
-            $user->password = bcrypt($request->password);
             $user->save();
 
             // Commit the transaction
@@ -212,7 +217,7 @@ class SRController extends Controller
             DB::rollback();
 
             // Handle the error and redirect with an error message
-            return redirect()->route('sr.edit',['dsr'=>$sr->id])
+            return redirect()->route('sr.edit',['sr'=>$sr->id])
                 ->withInput()
                 ->with('error', 'An error occurred while updating the SR: '.$e->getMessage());
         }
@@ -225,8 +230,15 @@ class SRController extends Controller
     public function destroy(Client $sr)
     {
         try {
-            // Delete the Client record
-            $sr->delete();
+            $saleOrder = SaleOrder::where('sr_id',$sr->id)->exists();
+            if($saleOrder){
+                return response()->json(['success'=>false,'message' => 'This SR cannot be deleted because it is associated with an existing Sale Order.'], Response::HTTP_OK);
+            }else{
+                User::where('client_id',$sr->id)->delete();
+                // Delete the Client record
+                $sr->delete();
+            }
+
             // Return a JSON success response
             return response()->json(['success'=>true,'message' => 'SR deleted successfully'], Response::HTTP_OK);
         } catch (\Exception $e) {
